Support an extra source files under similarly named directory.
We wish to support splitting a go "script" in to multiple files, as a
script grows. We also wish to allow multiple different scripts to live on
the PATH (e.g. all in /usr/local/bin) without interfering with each other.

This change copies and extra directory with a similar name to the original
source code and allows that to be used. e.g.:
./httpServe.go
./httpServe_/net/auth.go
./httpServe_/net/reply.go
./httpServe_/db/sql.go

with import "httpServe/httpServe_/net" in httpServe.go etc.

That way, it should be obvious on a server what files live with what script